requirements.txt
einops==0.8.0
transformers==4.43.3
huggingface-hub
diffusers==0.30.1
sentencepiece==0.2.0
gradio==5.22.0
b2sdk==1.20.0
requests
Pillow

--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.4.0
torchvision==0.19.0
